Update keybindings for okesl 0.10
Previous versions of okesl had foul language exposed to users and there
was a decision to get rid of this language. This updates the configs
supplied in this repository to work for okesl 0.10 by updating the foul
language with their new counterparts. Also some removed variables have
been removed from the configs